@@ -3,8 +3,8 @@
 namespace snrk {
 
 template<typename V>
-Value<V>::Value(V)
-    : m_data{std::make_shared<V>(new V)}
+Value<V>::Value(V v)
+    : m_data{std::make_shared<V>(new V(v))}
 {
 
 }
@@ -25,6 +25,14 @@ Gate<V>::Gate(type_t type, input_t input, value_t output)
 }
 
 template<typename V>
+Circut<V>::Circut(const input_t &inputX, const input_t &inputW)
+    : m_inputX{inputX}
+    , m_inputW{inputW}
+{
+
+}
+
+template<typename V>
 std::size_t Circut<V>::size() const
 {
     return m_gates.size();
@@ -40,6 +48,14 @@ template<typename V>
 std::size_t Circut<V>::degree() const
 {
     return 3 * size() + inputSize();
+}
+
+template<typename V>
+void Circut<V>::addGate(const gate_t &gate)
+{
+    /*todo: мютекс, если многопоточка*/
+
+    m_gates.push_back(gate);
 }
 
 }
